@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
   repositories {
-    jcenter()
     google()
+    mavenCentral()
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.5.3'
-    classpath 'org.sonarsource.scanner.gradle:sonarqube-gradle-plugin:2.8'
+    classpath 'com.android.tools.build:gradle:8.2.0'
+    // classpath 'org.sonarsource.scanner.gradle:sonarqube-gradle-plugin:2.8'
 
     // NOTE: Do not place your application dependencies here; they belong
     // in the individual module build.gradle files
@@ -15,8 +15,8 @@ buildscript {
 
 allprojects {
   repositories {
-    jcenter()
     google()
+    mavenCentral()
     maven { url 'https://jitpack.io' }
   }
 }
@@ -25,20 +25,16 @@ task clean(type: Delete) {
   delete rootProject.buildDir
 }
 
-
-
 // This block encapsulates custom properties and makes them available to all
 // modules in the project.
 ext {
   // The following are only a few examples of the types of properties you can define.
-  compileSdkVersion = 29
+  compileSdkVersion = 34
   // You can also create properties to specify versions for dependencies.
   // Having consistent versions between modules can avoid conflicts with behavior.
-  supportLibVersion = '28.0.0'
-  gmsFirebaseVersion = '16.0.6'
   minSdkVersion = 24
-  targetSdkVersion = 29
-  jUnitVersion = '4.12'
+  targetSdkVersion = 34
+  jUnitVersion = '4.13.2'
   applicationId = 'in.rahulja.groupingmessages'
   versionCode = 16
   versionName = '1.6'
